requirements.txt
osmnx==1.9.2
networkx==3.2
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
scikit-learn
numpy
shapely